backapp/requirements.txt
astroid==2.1.0
dj-database-url==0.5.0
Django==2.1.5
django-heroku==0.3.1
djangorestframework==3.9.1
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
pkg-resources==0.0.0
psycopg2==2.7.7
pylint==2.2.2
pytz==2018.9
six==1.12.0
typed-ast==1.2.0
whitenoise==4.1.2
wrapt==1.11.1